requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
alembic==1.0.9
click==7.0
dominate==2.3.5
flask-bootstrap==3.3.7.1
flask-login==0.4.1
flask-migrate==2.4.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
itsdangerous==1.1.0
jinja2==2.10.1
mako==1.0.9
markupsafe==1.1.1
pycrypto==2.6.1
pymysql==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
sqlalchemy==1.3.3
visitor==0.1.3
werkzeug==0.15.2
wtforms==2.2.1
